+++ Source/Core/HCAssertThat.h
@@ -154,34 +154,3 @@ OBJC_EXPORT void HC_assertWithTimeoutAndLocation(id testCase, NSTimeInterval tim
  */
 #define thatEventually HC_thatEventually
 #endif
-
-
-#define HC_assertWithTimeoutC(timeout, actualBlock, matcher)  \
-    HC_assertWithTimeoutAndLocation(nil, timeout, actualBlock, matcher, __FILE__, __LINE__)
-
-#ifdef HC_SHORTHAND
-/*!
- * @brief assertWithTimeoutC(timeout, actualBlock, matcher) -
- * Asserts that a value provided by a block will satisfy matcher within a given time, in a standalone function.
- * @param timeout Maximum time to wait for passing behavior, specified in seconds.
- * @param actualBlock A block providing the object to repeatedly evaluate as the actual value.
- * @param matcher The matcher to satisfy as the expected condition.
- * @discussion assertWithTimeoutC polls a value provided by a block to asynchronously satisfy the
- * matcher. The block is evaluated repeatedly for an actual value, which is passed to the matcher
- * for evaluation. If the matcher is not satisfied within the timeout, it is reported to the
- * @ref HCTestFailureReporterChain.
- *
- * An easy way of providing the <em>actualBlock</em> is to use the macro
- * <code>thatEventually(actual)</code>.
- *
- * Use assertWithTimeoutC in standalone C functions. It's designed to integrate with testing
- * frameworks where individual tests are executed as standalone functions. You may also want to
- * install a custom test failure handler for better reporting.
- *
- * @see thatEventually
- * @attribute Name Clash
- * In the event of a name clash, don't <code>#define HC_SHORTHAND</code> and use the synonym
- * HC_assertWithTimeoutC instead.
-*/
-#define assertWithTimeoutC HC_assertWithTimeoutC
-#endif