+++ Source/Tests/Core/AssertWithTimeout.m
@@ -11,13 +11,6 @@
 
 static NSTimeInterval const TIME_ERROR_MARGIN = 0.1f;
 
-static void standaloneAssertWithTimeout(NSTimeInterval timeout,
-        HCFutureValue actualBlock,
-        id <HCMatcher> matcher)
-{
-    assertWithTimeoutC(timeout, actualBlock, matcher);
-}
-
 
 @interface AssertWithTimeout : SenTestCase
 @end
@@ -90,18 +83,6 @@ static void standaloneAssertWithTimeout(NSTimeInterval timeout,
 
     STAssertTrue(waitTime > succeedTime, @"Expect assert to terminate after value is changed, but was %lf", waitTime);
     STAssertTrue(waitTime < maxTime, @"Expect assert to terminate before timeout, but was %lf", waitTime);
-}
-
-- (void)testStandaloneAssertWithTimeoutC_EventuallySatisfied_ShouldSucceed
-{
-    NSTimeInterval succeedTime = 0.2;
-    __block NSString *futureBar = @"foo";
-    dispatch_time_t popTime = dispatch_time(DISPATCH_TIME_NOW, (int64_t)(succeedTime * NSEC_PER_SEC));
-    dispatch_after(popTime, dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^(void) {
-        futureBar = @"bar";
-    });
-
-    standaloneAssertWithTimeout(1, thatEventually(futureBar), equalTo(@"bar"));
 }
 
 - (void)testDeprecatedAssertThatAfter_WithTimeoutGreaterThanZeroShouldSucceedNotImmediatelyButBeforeTimeout